==> notebooks/01_data_loading.ipynb <==
### Chargement de données

##### 1. Charger le dataset SetFit/ag_news depuis Hugging Face avec la bibliothèque datasets.

from datasets import load_dataset

dataset = load_dataset("SetFit/ag_news")


##### 2. Afficher le contenu du dataset

print(dataset)

##### 3. Convertir les données en DataFrame pandas.

import pandas as pd

data_train = pd.DataFrame(dataset["train"])
data_test = pd.DataFrame(dataset["test"])


##### 4. Sauvegarder les données chargées

data_train.to_csv('../data/raw/train.csv', index=False)
data_test.to_csv('../data/raw/test.csv', index=False)
