==> notebooks/03_preprocessing.ipynb <==
### Prétraitement des textes

##### 1. Charger les données 

import pandas as pd

data_train = pd.read_csv('../data/raw/train.csv')
data_test = pd.read_csv('../data/raw/test.csv')

##### 2. Normaliser les textes

- Mettre en minuscules

data_train['text'] = data_train['text'].str.lower()
data_test['text'] = data_test['text'].str.lower()

- Supprimer les chiffres et la ponctuation

import re

data_train['text_cleaned'] = data_train['text'].apply(lambda x: re.sub(r'[^\w\s\.]', '', x))
data_test['text_cleaned'] = data_test['text'].apply(lambda x: re.sub(r'[^\w\s\.]', '', x))

- Supprimer les points (en gardant les URLs)

data_train['text_cleaned'] = data_train['text_cleaned'].apply(lambda x: re.sub(r'\. ', ' ', x))
data_test['text_cleaned'] = data_test['text_cleaned'].apply(lambda x: re.sub(r'\. ', ' ', x))

- Supprimer les espaces multiples

import re

data_train['text_cleaned'] = data_train['text_cleaned'].apply(lambda x: re.sub(r'\s+', ' ', x).strip())
data_test['text_cleaned'] = data_test['text_cleaned'].apply(lambda x: re.sub(r'\s+', ' ', x).strip())

##### 3. Supprimer les stopwords.

from nltk.corpus import stopwords

stop_words = set(stopwords.words('english'))

data_train['tokens'] = data_train['text_cleaned'].apply(lambda x: [w for w in x.split() if w not in stop_words])
data_test['tokens'] = data_test['text_cleaned'].apply(lambda x: [w for w in x.split() if w not in stop_words])


##### 4. Vérifier et valider la qualité finale des textes prétraités.

data_train['text_final'] = data_train['tokens'].apply(lambda x: " ".join(x))
data_test['text_final'] = data_test['tokens'].apply(lambda x: " ".join(x))

- Train

data_train['text_final'].head(20)

- Test

data_test['text_final'].head(20)

##### 5. Sauvegarder les changements

data_train[['text_final', 'label', 'label_text']].to_csv('../data/processed/train_text_cleaned.csv', index=False)
data_test[['text_final', 'label', 'label_text']].to_csv('../data/processed/test_text_cleaned.csv', index=False)

##### 6. Stocker les noms des classes

classes = data_train[~data_train[['label', 'label_text']].duplicated()][['label', 'label_text']]

classes.to_csv('../data/processed/classes.csv', index=False)
