==> notebooks/04_embeddings.ipynb <==
### Génération d’embeddings

##### 1. Le choix de modèles génération d’embeddings

Parmi les meilleurs modèles de génération d’embeddings, on trouve :

| Modèle              | Qualité         | Multilingue       | Vitesse |
| ------------------- | --------------- | ----------------- | ------- |
| BGE-large-m3        | Excellent       | Oui               | Lent |
| E5-large-v2         | Excellent       | EN only           | Moyen |
| Multilingual-E5     | Bien            | Oui               | Moyen |
| GTE-large           | Bien            | EN only           | Rapide |
| MiniLM-L12-v2       | qualité faible  | Oui               | Très rapide |


On va utiliser `E5-large-v2` pour trouver un équilibre entre la performance et la vitesse d’entraînement.

##### 2. Prétraitement des textes pour le modéle `E5-large-v2`

Le modèle `E5-large-v2` a été entraîné sur du texte naturel (phrases réelles, majuscules, ponctuation, etc.), donc il faut garder le texte aussi proche que possible de sa forme originale,
tout en nettoyant les erreurs mineures ou les incohérences.

il est recommander d'éviter de:
- Supprimer les stopwords: Le modèle comprend leur rôle sémantique.
- Lemmatisation ou stemming: Cela modifie la structure linguistique et le sens.
- Supprimer la ponctuation: La ponctuation aide à comprendre le contexte.
- Tout mettre en minuscules (Sauf si le texte contient des majuscules aléatoires ou tout en majuscules.): Peut faire perdre des distinctions importantes (Apple ≠ apple, US ≠ us).
- Supprimer les chiffres: Les nombres ont souvent une signification utile (dates, montants…).

Il est Recommander (et parfois obligatoire) de:
- Supprimer les espaces inutiles: Nettoyer les débuts/fin de lignes et les espaces multiples.
- Gérer les valeurs manquantes: Remplacer les `NaN` ou textes vides par une chaîne vide.
- Ajouter le préfixe (Obligatoire): **"passage: "** pour les documents, **"query: "** pour les requêtes.
- Tronquer les textes très longs: Si un texte dépasse la limite (512 tokens environ), on peut le couper.

import pandas as pd

data_train = pd.read_csv('../data/raw/train.csv')
data_test = pd.read_csv('../data/raw/test.csv')


data_train["text"] = data_train["text"].fillna("").str.strip()
data_test["text"] = data_test["text"].fillna("").str.strip()


data_train["text"] = data_train["text"].str.replace(r"\s+", " ", regex=True)
data_test["text"] = data_test["text"].str.replace(r"\s+", " ", regex=True)


data_train["text_prefixed"] = "passage: " + data_train["text"]
data_test["text_prefixed"] = "passage: " + data_test["text"]

- Sauvegarder les modifications

data_train.to_csv('../data/processed/preprocessed_train_text.csv', index=False)
data_test.to_csv('../data/processed/preprocessed_test_text.csv', index=False)

##### 3. Charger le modèle `e5-large-v2` avec **Sentence Transformers**

from sentence_transformers import SentenceTransformer

model = SentenceTransformer("intfloat/e5-large-v2")

##### 4. Générer les embeddings pour toutes les données.

- Train

data_train["embedding"] = data_train["text_prefixed"].apply(
    lambda x: model.encode(x, normalize_embeddings=True)
)

- Test

data_test["embedding"] = data_test["text_prefixed"].apply(
    lambda x: model.encode(x, normalize_embeddings=True)
)

Malheureusement, la génération des embeddings prend beaucoup de temps (plus de 400 minutes pour les données d'entraînement sans se terminer). Nous allons donc utiliser un autre modèle qui est plus rapide.

##### 5. Générer les embeddings avec le modèle `paraphrase-multilingual-MiniLM-L12-v2`

Prétraiter le texte:

- Supprimer les espaces inutiles: Nettoyer les débuts/fin de lignes et les espaces multiples.
- Gérer les valeurs manquantes: Remplacer les `NaN` ou textes vides par une chaîne vide.

from sentence_transformers import SentenceTransformer

model = SentenceTransformer("paraphrase-multilingual-MiniLM-L12-v2")

- Train

data_train["text"] = data_train["text"].fillna("").str.strip()
data_train["text"] = data_train["text"].str.replace(r"\s+", " ", regex=True)

texts = data_train["text"].tolist()

train_embeddings = model.encode(
    texts,
    batch_size=32,
    show_progress_bar=True,
    convert_to_numpy=True,
    normalize_embeddings=True
)

import numpy as np
import os 

os.makedirs('../data/embeddings', exist_ok=True)

np.save("../data/embeddings/train_embeddings_minilm.npy", train_embeddings)

- Test

data_test["text"] = data_test["text"].fillna("").str.strip()
data_test["text"] = data_test["text"].str.replace(r"\s+", " ", regex=True)

texts = data_test["text"].tolist()

test_embeddings = model.encode(
    texts,
    batch_size=32,
    show_progress_bar=True,
    convert_to_numpy=True,
    normalize_embeddings=True
)

import numpy as np
import os 

os.makedirs('../data/embeddings', exist_ok=True)

np.save("../data/embeddings/test_embeddings_minilm.npy", test_embeddings)

- Sauvegardez le modéle

import joblib as jb

jb.dump(model, '../models/minilm_embedding_model.pkl')

##### 6. Vérification des embeddings générés

print(train_embeddings.shape, '\n')
print(train_embeddings)

print(test_embeddings.shape, '\n')
print(test_embeddings)
