==> notebooks/05_chromadb_storage.ipynb <==
### Stockage de données

##### 1. Chargement de données

import pandas as pd

data_train = pd.read_csv('../data/processed/preprocessed_train_text.csv')
data_test = pd.read_csv('../data/processed/preprocessed_test_text.csv')


##### 2. Extraire et sauvegarder les métadonnées nécessaires (label, identifiant).

train_metadata = data_train[["label", "label_text"]].copy()
test_metadata = data_test[["label", "label_text"]].copy()

train_metadata["id"] = range(len(train_metadata))
test_metadata["id"] = range(len(test_metadata))

- Sauvegarder

import os

os.makedirs('../data/metadata', exist_ok=True)

train_metadata.to_csv("../data/metadata/train_metadata.csv", index=False)
test_metadata.to_csv("../data/metadata/test_metadata.csv", index=False)

##### 2. Créer les collections ChromaDB

- Initialisation

import chromadb

client = chromadb.PersistentClient(path='../data/chromaDB')

- Pour les données d’entraînement.

train_collection = client.create_collection(
    name="train_collection",
    metadata={"description": "Embeddings pour données d'entraînement"}
)

- Pour les données de test.

test_collection = client.create_collection(
    name="test_collection",
    metadata={"description": "Embeddings pour données de test"}
)

##### 3. Insérer les embeddings et métadonnées dans les collections.

- la colonne `ids` doit toujours être une liste de chaînes de caractères (str), pas des entiers.
- `ids` doit être unique pour chaque document.
- `embeddings` doit être une liste de listes, pas un array Numpy.
- Diviser les données en plus petits lots avant de les ajouter à la collection. (ValueError si les lignes se dépasse 5461)

1. Charger les embeddings

import numpy as np

train_embeddings = np.load('../data/embeddings/train_embeddings_minilm.npy')
test_embeddings = np.load('../data/embeddings/test_embeddings_minilm.npy')


2. Train

batch_size = 5000
n = len(train_embeddings)

for i in range(0, n, batch_size):
    end = min(i + batch_size, n)
    
    batch_ids = [str(j) for j in list(train_metadata['id'][i: end])]
    batch_embeddings = train_embeddings[i:end].tolist()
    batch_metadatas = [
        {
            "label": int(train_metadata['label'][j]),
            "label_text": str(train_metadata['label_text'][j])
        } for j in range(i, end)
    ]
    
    train_collection.add(
        ids=batch_ids,
        embeddings=batch_embeddings,
        metadatas=batch_metadatas,
    )
    
    print(f"Inserted {end} / {n}")

3. Test

batch_size = 5000
n = len(test_embeddings)

for i in range(0, n, batch_size):
    end = min(i + batch_size, n)
    
    batch_ids = [str(j) for j in list(test_metadata['id'][i: end])]
    batch_embeddings = test_embeddings[i:end].tolist()
    batch_metadatas = [
        {
            "label": int(test_metadata['label'][j]),
            "label_text": str(test_metadata['label_text'][j])
        } for j in range(i, end)
    ]
    
    test_collection.add(
        ids=batch_ids,
        embeddings=batch_embeddings,
        metadatas=batch_metadatas
    )
    
    print(f"Inserted {end} / {n}")


##### 4. Vérifier l’intégrité des données stockées

print("Le nombre total de collection d'entrainement:", train_collection.count())
print("Le nombre total de collection de test:", test_collection.count())

results = train_collection.query(
    query_embeddings=[train_embeddings[0].tolist()],
    n_results=1,
    include=['embeddings', 'metadatas']
)

print(results.keys(), '\n')

results
