==> notebooks/06_training.ipynb <==
### Entrainement de modéles

##### 1. Récupérer les embeddings depuis ChromaDB.

import chromadb

client = chromadb.PersistentClient(path='../data/chromaDB')

train_collection = client.get_collection(name='train_collection')
test_collection = client.get_collection(name='test_collection')

print(train_collection.count())
print(test_collection.count())

##### 2. Extraction de labels et d'embeddings

all_metadatas = train_collection.get(include=["metadatas", "embeddings"])

labels = [metadata['label'] for metadata in all_metadatas['metadatas']]

embeddings = [emb for emb in all_metadatas['embeddings']]

print(len(labels))
print(len(embeddings))

##### 3. Entraîner différents modèles ML sur les embeddings.

- Préparer les données (features & target)

import numpy as np

X_train = np.array(embeddings)
y_train = np.array(labels)

- Importer des modèles ML

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from xgboost import XGBClassifier

- Logistic Regression

lr_model = LogisticRegression()
lr_model.fit(X_train, y_train)

- Random Forest

rf_model = RandomForestClassifier()
rf_model.fit(X_train, y_train)

- XGBoost

xgb_model = XGBClassifier()

xgb_model.fit(embeddings, y_train)


- SVM

svm_model = SVC()
svm_model.fit(X_train, y_train)

- K-Nearest Neighbors (KNN)

knn_model = KNeighborsClassifier()
knn_model.fit(X_train, y_train)

##### 4. Sauvegarder les modéles

import joblib as hafid

hafid.dump(lr_model, '../models/lr_model.pkl')
hafid.dump(rf_model, '../models/rf_model.pkl')
hafid.dump(xgb_model, '../models/xgb_model.pkl')
hafid.dump(svm_model, '../models/svm_model.pkl')
hafid.dump(knn_model, '../models/knn_model.pkl')
